+++ Runtime/UniShaderUrpUtility/Defines/IgnoreProjectorValue.cs
@@ -1,0 +1,18 @@
+﻿// ----------------------------------------------------------------------
+// @Namespace : UniUrpShader
+// @Class     : IgnoreProjectorValue
+// ----------------------------------------------------------------------
+namespace UniUrpShader
+{
+    /// <summary>
+    /// Ignore Projector Value
+    /// </summary>
+    public class IgnoreProjectorValue
+    {
+        /// <summary>True</summary>
+        public const string True = "True";
+
+        /// <summary>False</summary>
+        public const string False = "False";
+    }
+}