+++ Runtime/UniShaderUrpUtility/Defines/RenderPipelineValue.cs
@@ -1,0 +1,15 @@
+﻿// ----------------------------------------------------------------------
+// @Namespace : UniUrpShader
+// @Class     : RenderPipelineValue
+// ----------------------------------------------------------------------
+namespace UniUrpShader
+{
+    /// <summary>
+    /// Render Pipeline Value
+    /// </summary>
+    public class RenderPipelineValue
+    {
+        /// <summary>Universal Pipeline</summary>
+        public const string UniversalPipeline = "UniversalPipeline";
+    }
+}