+++ Runtime/UniShaderUrpUtility/Defines/RenderTypeValue.cs
@@ -1,0 +1,21 @@
+﻿// ----------------------------------------------------------------------
+// @Namespace : UniUrpShader
+// @Class     : RenderTypeValue
+// ----------------------------------------------------------------------
+namespace UniUrpShader
+{
+    /// <summary>
+    /// Render Type Value
+    /// </summary>
+    public class RenderTypeValue
+    {
+        /// <summary>Opaque</summary>
+        public const string Opaque = "Opaque";
+
+        /// <summary>Transparent</summary>
+        public const string Transparent = "Transparent";
+
+        /// <summary>Transparent Cutout</summary>
+        public const string TransparentCutout = "TransparentCutout";
+    }
+}