+++ Runtime/UniShaderUrpUtility/Defines/Tag.cs
@@ -9,5 +9,16 @@ namespace UniUrpShader
     /// </summary>
     public class Tag
     {
+        /// <summary>Render Type</summary>
+        public const string RenderType = "RenderType";
+
+        /// <summary>Render Pipeline</summary>
+        public const string RenderPipeline = "RenderPipeline";
+
+        /// <summary>Universal Material Type</summary>
+        public const string UniversalMaterialType = "UniversalMaterialType";
+
+        /// <summary>Ignore Projector</summary>
+        public const string IgnoreProjector = "IgnoreProjector";
     }
 }