+++ Runtime/UniShaderUrpUtility/Defines/UniversalMaterialTypeValue.cs
@@ -1,0 +1,24 @@
+﻿// ----------------------------------------------------------------------
+// @Namespace : UniUrpShader
+// @Class     : UniversalMaterialTypeValue
+// ----------------------------------------------------------------------
+namespace UniUrpShader
+{
+    /// <summary>
+    /// Universal Material Type Value
+    /// </summary>
+    public class UniversalMaterialTypeValue
+    {
+        /// <summary>Lit</summary>
+        public const string Lit = "Lit";
+
+        /// <summary>Simple Lit</summary>
+        public const string SimpleLit = "SimpleLit";
+
+        /// <summary>Complex Lit</summary>
+        public const string ComplexLit = "ComplexLit";
+
+        /// <summary>Unlit</summary>
+        public const string Unlit = "Unlit";
+    }
+}