+++ Runtime/UniShaderUrpUtility/Enums/BlendMode.cs
@@ -7,11 +7,13 @@ namespace UniUrpShader
     /// <summary>Blend Mode</summary>
     public enum BlendMode
     {
-        /// <summary></summary>
+        /// <summary>Alpha blend mode.</summary>
         Alpha = 0,
-        /// <summary></summary>
-        Additive = 1,
-        /// <summary></summary>
-        Premultiply = 4,
+        /// <summary>Premultiply blend mode.</summary>
+        Premultiply = 1,
+        /// <summary>Additive blend mode.</summary>
+        Additive = 2,
+        /// <summary>Multiply blend mode.</summary>
+        Multiply = 3,
     }
 }