+++ Runtime/UniShaderUrpUtility/Enums/SmoothnessSource.cs
@@ -7,9 +7,9 @@ namespace UniUrpShader
     /// <summary>Smoothness Source</summary>
     public enum SmoothnessSource
     {
-        /// <summary></summary>
+        /// <summary>Use this when smoothness is stored in the alpha channel of the specular map.</summary>
         SpecularAlpha = 0,
-        /// <summary></summary>
+        /// <summary>Use this when smoothness is stored in the alpha channel of the base map.</summary>
         BaseAlpha = 1,
     }
 }